requirements.txt
numpy==1.22.4
sounddevice==0.4.3
python-dotenv==1.0.1
playsound==1.3.0
sentence-transformers==3.2.1
python-Levenshtein==0.25.1
torch==2.4.1
torchvision==0.19.1
opencv-python==4.10.0.84
onnxruntime==1.12.0
speechrecognition==3.10.4
scipy==1.7.3
PyObjC==10.3.2